+++ resources/assets/js/components/movies/index.js
